requirements.txt
scipy==0.19.0
psycopg2==2.7.1
requests==2.18.3
numpy==1.13.0
Shapely==1.5.17.post1
pandas==0.20.2
plotly==2.0.13
polyline==1.3.2
matplotlib==2.0.2
geopandas==0.2.1
ipython==6.2.0
python_igraph==0.7.1.post6
